tidy-up

+++ src/main/kotlin/kotlin4clojure/easy/fibonacci_sequence/FibonacciSequence.kt
@@ -1,5 +1,5 @@
 package kotlin4clojure.easy.fibonacci_sequence
 
-fun fibonacciSequence(input: Int): List<Int> {
+fun fibonacciSequence(input: Int): List<Long> {
     TODO("Add your solution here")
 }
